+++ src/main/java/nl/dtls/fairdatapoint/entity/resource/ResourceDefinition.java
@@ -37,6 +37,7 @@ import java.util.List;
 @Getter
 @Setter
 @EqualsAndHashCode
+@Builder(toBuilder = true)
 public class ResourceDefinition {
 
     @Id